% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{add_set_data}
\alias{add_set_data}
\title{Add data to a set}
\usage{
add_set_data(model, set_name, members)
}
\arguments{
\item{model}{Multimod model}

\item{set_name}{Name of set}

\item{members}{Character vector of set members}
}
\value{
Modified model
}
\description{
Add data to a set
}
