% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{annotate_brackets}
\alias{annotate_brackets}
\title{Annotate expression tree with bracket requirements}
\usage{
annotate_brackets(expr, parent = NULL, side = NULL)
}
\description{
Annotate expression tree with bracket requirements
}
