% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_character.R
\name{as.character.expression}
\alias{as.character.expression}
\title{Convert AST expression to character}
\usage{
\method{as.character}{expression}(x, brackets = NULL, max_char = 50, ...)
}
\arguments{
\item{x}{An expression object}

\item{brackets}{Bracket style to use}

\item{max_char}{Maximum characters before truncation}

\item{...}{Additional arguments}
}
\description{
Convert AST expression to character
}
