% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast-classes.R
\name{as_dims}
\alias{as_dims}
\title{Method for converting various objects to \code{dims}}
\usage{
as_dims(x, ...)
}
\arguments{
\item{x}{The object to be converted.}
}
\description{
This function is a generic method for converting various objects
to \code{dims} objects.
}
