% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gams.R
\name{as_gams.equation}
\alias{as_gams.equation}
\title{Convert equation to GAMS syntax}
\usage{
as_gams.equation(eqn)
}
\arguments{
\item{eqn}{A \code{equation} object.}
}
\value{
A character string with valid GAMS syntax.
}
\description{
Render a \code{equation} object as a GAMS equation string.
}
