% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_gams.R
\name{as_gams}
\alias{as_gams}
\alias{as_gams.func}
\title{Convert multimod object to GAMS syntax}
\usage{
as_gams(x, ...)

\method{as_gams}{func}(x, ...)
}
\arguments{
\item{x}{A \code{multimod} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A character string with valid GAMS syntax.
}
\description{
Render a \code{multimod} object as a GAMS code string.
}
