% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_gmpl.R
\name{as_gmpl}
\alias{as_gmpl}
\alias{as_gmpl.call}
\alias{as_gmpl.func}
\title{Convert multimod object to GMPL/MathProg syntax}
\usage{
as_gmpl(x, ...)

\method{as_gmpl}{call}(x, ...)

\method{as_gmpl}{func}(x, ...)
}
\arguments{
\item{x}{A \code{multimod} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A character string with valid GMPL syntax.
}
\description{
Render a \code{multimod} object as a GMPL code string.
}
