% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_gmpl.R
\name{as_gmpl_sum_prod}
\alias{as_gmpl_sum_prod}
\title{Convert GAMS-style sum/prod to GMPL binding syntax}
\usage{
as_gmpl_sum_prod(x, ...)
}
\arguments{
\item{x}{A func AST node representing sum or prod}

\item{...}{Additional arguments passed to as_gmpl}
}
\value{
GMPL sum/prod expression with proper binding
}
\description{
Convert GAMS-style sum/prod to GMPL binding syntax
}
\keyword{internal}
