% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_jump.R
\name{as_jump}
\alias{as_jump}
\title{Convert multimod object to Julia/JuMP syntax}
\usage{
as_jump(x, ...)
}
\arguments{
\item{x}{A \code{multimod} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A character string with valid Julia/JuMP syntax.
}
\description{
Render a \code{multimod} object as a JuMP code string.
}
