% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_multimod.R
\name{as_multimod}
\alias{as_multimod}
\title{Convert an object to a multimod structure}
\usage{
as_multimod(x, ...)
}
\arguments{
\item{x}{An object to convert.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A multimod object.
}
\description{
This is a generic function that converts an object to a multimod-compatible format.
}
