% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_r.R
\name{as_r.ast_formula}
\alias{as_r.ast_formula}
\title{Convert ast_formula to R Code}
\usage{
\method{as_r}{ast_formula}(x, ...)
}
\arguments{
\item{x}{ast_formula object}

\item{...}{Additional arguments passed to as_r for the wrapped expression}
}
\value{
R code string
}
\description{
Convert ast_formula to R Code
}
