% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_r.R
\name{as_r}
\alias{as_r}
\title{Convert AST to R Code}
\usage{
as_r(x, ...)
}
\arguments{
\item{x}{An AST node (expression, symbol, constant, etc.)}

\item{...}{Additional arguments passed to methods}
}
\value{
A character string containing R code
}
\description{
Converts AST nodes to executable R code. This is the foundation for
creating R functions from AST expressions.
}
