% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_vizNetwork.R
\name{as_visNetwork}
\alias{as_visNetwork}
\alias{as_visNetwork.ast}
\title{Convert a multimod ast or equation to a visNetwork object}
\usage{
as_visNetwork(x, ...)

\method{as_visNetwork}{ast}(
  x,
  main = x$name,
  submain = x$desc,
  alias_map = NULL,
  show_dims = TRUE
)
}
\arguments{
\item{x}{A multimod object}
}
\description{
Convert a multimod ast or equation to a visNetwork object
}
