% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast-classes.R
\name{ast_call}
\alias{ast_call}
\title{Create a function call AST node}
\usage{
ast_call(name, args = list())
}
\arguments{
\item{name}{Character string, function name}

\item{args}{List of AST nodes representing arguments}
}
\value{
An object of class \code{ast} and \code{call}
}
\description{
Constructs an AST node representing a function call with arguments.
This is similar to ast_func but specifically for function calls in expressions.
}
