% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast-classes.R
\name{ast_condition}
\alias{ast_condition}
\title{Construct a conditional (dollar) expression node for multimod AST}
\usage{
ast_condition(condition, then)
}
\arguments{
\item{condition}{The condition to check (right-hand side of \code{$}). Must be an AST node.}

\item{then}{The expression to evaluate if the condition is true (usually the left-hand side).}
}
\value{
An object of class \code{multimod_ast} and subclass \code{ast_condition}
}
\description{
This function constructs a conditional expression node of type \code{"condition"},
representing GAMS-style conditional terms using the \code{$} operator.
}
\examples{
ast_condition(
  condition = ast_symbol("i_active(i)"),
  then = ast_variable("x", c("i"))
)
}
