% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast-classes.R
\name{ast_constant}
\alias{ast_constant}
\title{Create a constant AST node}
\usage{
ast_constant(value)
}
\arguments{
\item{value}{A numeric or character constant.}
}
\value{
An \code{ast_constant} S3 object (subclass of \code{multimod_ast}).
}
\description{
Create a constant AST node
}
