% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast-classes.R
\name{ast_mapping}
\alias{ast_mapping}
\title{Create a mapping AST node}
\usage{
ast_mapping(name, dims = ast_dims(), domain = NULL, ...)
}
\arguments{
\item{name}{A character string representing the mapping name.}

\item{dims}{An \code{ast_dims} object representing the dimensions of the mapping.}
}
\value{
An \code{ast_mapping} S3 object (subclass of \code{multimod_ast}).
}
\description{
Constructs a mapping node representing a mapping or set of indices.
}
