% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast-classes.R
\name{ast_parse_expr}
\alias{ast_parse_expr}
\title{Language-aware AST expression parser}
\usage{
ast_parse_expr(expr, symbols = NULL, language = NULL, context = NULL, ...)
}
\description{
Language-aware AST expression parser
}
\keyword{internal}
