% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast-classes.R
\name{ast_prod}
\alias{ast_prod}
\title{Create a product AST node}
\usage{
ast_prod(index, value, domain = NULL)
}
\arguments{
\item{index}{An AST node representing the index variable(s) (e.g., \code{ast_dims("t")}).}

\item{value}{An AST node representing the expression to be summed.}

\item{domain}{An optional AST node representing a domain condition.
This can be a mapping, a logical condition (e.g., from a \code{$}-filter), or a parameter.
Use \code{NULL} if there is no restriction.}
}
\value{
An object of class \code{ast} and \code{prod}.
}
\description{
Constructs an abstract syntax tree (AST) node representing a product over an
index. The index is typically a \code{set}, \code{dims} or \code{when} object if
filtering is applied to the index.
}
