% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast-classes.R
\name{ast_setmax}
\alias{ast_setmax}
\title{Create a set maximum AST node}
\usage{
ast_setmax(index, value)
}
\arguments{
\item{index}{An AST node representing the index set}

\item{value}{An AST node representing the value expression}
}
\value{
An object of class \code{ast} and \code{setmax}
}
\description{
Constructs an AST node representing a maximum over a set index.
This is distinct from numeric max() functions.
}
