% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast-classes.R
\name{ast_setmin}
\alias{ast_setmin}
\title{Create a set minimum AST node}
\usage{
ast_setmin(index, value)
}
\arguments{
\item{index}{An AST node representing the index set}

\item{value}{An AST node representing the value expression}
}
\value{
An object of class \code{ast} and \code{setmin}
}
\description{
Constructs an AST node representing a minimum over a set index.
This is distinct from numeric min() functions.
}
