% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast-classes.R
\name{ast_symbol}
\alias{ast_symbol}
\title{Create a symbol AST node (for unclassified identifiers)}
\usage{
ast_symbol(name, ...)
}
\arguments{
\item{name}{A character string representing the symbol.}
}
\value{
An \code{symbol} S3 object (subclass of \code{ast}).
}
\description{
Create a symbol AST node (for unclassified identifiers)
}
