% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast-classes.R
\name{ast_type}
\alias{ast_type}
\title{Get the class of an AST node}
\usage{
ast_type(x)
}
\arguments{
\item{x}{An AST node object.}
}
\value{
A character string representing the class of the AST node.
}
\description{
This function retrieves the class of an AST node.
}
