% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{begin_math_env}
\alias{begin_math_env}
\title{Remove mapping nodes from an AST (including nested expressions)}
\usage{
begin_math_env(envs)
}
\arguments{
\item{ast}{A multimod AST object}
}
\value{
A cleaned AST object with mappings removed (NULL if fully removed)
}
\description{
This function traverses a multimod AST and removes all mapping nodes,
including those inside conditions and logical/arithmetic expressions.
}
