% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_character.R
\name{brackets_pair}
\alias{brackets_pair}
\title{Get a pair of brackets}
\usage{
brackets_pair(x)
}
\arguments{
\item{x}{A character string or vector of length 1 or 2.
If the length is 1, it will return a pair of brackets based on the
character. If the length is 2, it will return the two characters as a vector.}
}
\value{
A character vector of length 2 containing the opening and closing brackets.
}
\description{
This function returns a pair of brackets based on the input character.
}
