% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_jump.R
\name{build_product_ast}
\alias{build_product_ast}
\title{Build a nested product AST from a list of factor nodes}
\usage{
build_product_ast(factors)
}
\description{
Build a nested product AST from a list of factor nodes
}
\keyword{internal}
