% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimod.R
\name{build_symbols_list}
\alias{build_symbols_list}
\title{Build a list of symbols from model_info}
\usage{
build_symbols_list(model_info)
}
\arguments{
\item{model_info}{model_structure object (as returned by \code{read_gams()})}
}
\value{
A list containing sets, mappings, parameters, and variables.
}
\description{
Build a list of symbols from model_info
}
