% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_jump.R
\name{collect_product_factors}
\alias{collect_product_factors}
\title{Flatten a chain of multiplicative expressions into individual factors}
\usage{
collect_product_factors(node)
}
\description{
Flatten a chain of multiplicative expressions into individual factors
}
\keyword{internal}
