% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{collect_scenario_parameter_data}
\alias{collect_scenario_parameter_data}
\title{Collect scenario parameter data into memory}
\usage{
collect_scenario_parameter_data(ert_param)
}
\arguments{
\item{ert_param}{energyRt parameter object}
}
\value{
data.frame with parameter values (may be empty)
}
\description{
Collect scenario parameter data into memory
}
\keyword{internal}
