% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{collect_trim_stats}
\alias{collect_trim_stats}
\title{Collect trimming statistics}
\usage{
collect_trim_stats(model)
}
\arguments{
\item{model}{A multimod model object}
}
\value{
List with trim statistics
}
\description{
Collect trimming statistics
}
\keyword{internal}
