% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_ast_simple}
\alias{compare_ast_simple}
\title{Compare AST objects using compare package with custom handling}
\usage{
compare_ast_simple(ast1, ast2)
}
\arguments{
\item{ast1}{First AST object}

\item{ast2}{Second AST object}
}
\value{
Result from compare::compare()
}
\description{
Alternative comparison using the compare package. Simpler but may be
more verbose for complex AST structures.
}
