% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_jump.R
\name{convert_filter_with_haskey}
\alias{convert_filter_with_haskey}
\title{Convert filter expression with haskey checks for parameters (for sum filters)}
\usage{
convert_filter_with_haskey(expr, model, ...)
}
\description{
Convert filter expression with haskey checks for parameters (for sum filters)
}
\keyword{internal}
