% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{create_link_log_entry}
\alias{create_link_log_entry}
\title{Create log entry for parameter/mapping link}
\usage{
create_link_log_entry(
  multimod_name,
  scenario_name,
  type,
  status,
  ert_param,
  inMemory
)
}
\description{
Create log entry for parameter/mapping link
}
\keyword{internal}
