% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_parameters.R
\name{default_index_aliases}
\alias{default_index_aliases}
\title{Generate default iterator aliases from dimension names}
\usage{
default_index_aliases(dim_names)
}
\description{
Generate default iterator aliases from dimension names
}
\keyword{internal}
