% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_parameters.R
\name{dim_entry_to_name}
\alias{dim_entry_to_name}
\title{Convert a single dimension entry to its name}
\usage{
dim_entry_to_name(entry)
}
\description{
Convert a single dimension entry to its name
}
\keyword{internal}
