% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_jump.R
\name{distribute_sum_product}
\alias{distribute_sum_product}
\title{Inject multiplicative factors into the body of a sum node}
\usage{
distribute_sum_product(expr_node)
}
\description{
Inject multiplicative factors into the body of a sum node
}
\keyword{internal}
