% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_jump.R
\name{.extract_set_members}
\alias{.extract_set_members}
\title{Extract set members from model data (parameters/mappings)}
\usage{
.extract_set_members(model, set_name)
}
\arguments{
\item{model}{multimod object}

\item{set_name}{Name of the set to extract members for}
}
\value{
Character vector of unique set members
}
\description{
Extract set members from model data (parameters/mappings)
}
\keyword{internal}
