% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_jump.R
\name{.generate_data_jl}
\alias{.generate_data_jl}
\title{Generate data.jl file that loads all data upfront (energyRt.jl compatible)}
\usage{
.generate_data_jl(
  model_dir,
  solver_dir,
  model = NULL,
  use_folded = FALSE,
  shifts_needed = NULL,
  export_data = FALSE,
  data_mode = c("external", "embedded")
)
}
\description{
Generate data.jl file that loads all data upfront (energyRt.jl compatible)
}
\keyword{internal}
