% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_storage.R
\name{.save_dataframe}
\alias{.save_dataframe}
\title{Internal: Save dataframe to disk}
\usage{
.save_dataframe(
  data,
  path,
  format,
  compression = "zstd",
  compression_level = 15,
  verbose = FALSE,
  label = NULL
)
}
\description{
Internal: Save dataframe to disk
}
\keyword{internal}
