% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_energyrt.R
\name{en_parse_domain_hint}
\alias{en_parse_domain_hint}
\title{Parse a single *@ domain hint comment}
\usage{
en_parse_domain_hint(hint, dims)
}
\arguments{
\item{hint}{Comment line like "*@ mTechNew(tech,region,year)"}

\item{dims}{Variable dimensions from declaration}
}
\value{
Domain mapping name or NULL
}
\description{
Parse a single *@ domain hint comment
}
\keyword{internal}
