% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_parameters.R
\name{ensure_parameter_defval}
\alias{ensure_parameter_defval}
\title{Ensure parameters retain a scalar default after evaluation}
\usage{
ensure_parameter_defval(param)
}
\description{
Ensure parameters retain a scalar default after evaluation
}
\keyword{internal}
