% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_parameters.R
\name{evaluate_default_expr}
\alias{evaluate_default_expr}
\title{Evaluate default expression AST}
\usage{
evaluate_default_expr(defval_ast, all_data, model, verbose = FALSE)
}
\description{
Evaluate default expression AST
}
\keyword{internal}
