% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_parameters.R
\name{evaluate_formula}
\alias{evaluate_formula}
\title{Evaluate formula AST over all dimension combinations}
\usage{
evaluate_formula(
  formula_ast,
  dims,
  dims_index_aliases,
  all_data,
  model,
  verbose = FALSE
)
}
\description{
Evaluate formula AST over all dimension combinations
}
\keyword{internal}
