% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glpk_solver.R
\name{execute_glpkAPI}
\alias{execute_glpkAPI}
\title{Solve GLPK model using glpkAPI}
\usage{
execute_glpkAPI(
  model_dir,
  mod_file = NULL,
  dat_file = NULL,
  output_dir = NULL,
  verbose = TRUE,
  save_solution = FALSE,
  export_vars = FALSE,
  export_mps = FALSE
)
}
\arguments{
\item{model_dir}{Path to directory containing .mod and .dat files}

\item{mod_file}{Name of .mod file (default: searches for *.mod)}

\item{dat_file}{Name of .dat file (default: searches for *.dat)}

\item{output_dir}{Directory for solution output (default: model_dir/output)}

\item{verbose}{Logical, print solver output (default: TRUE)}

\item{save_solution}{Logical, save solution to CSV files (default: FALSE)}

\item{export_mps}{Logical, export model to MPS format (default: FALSE)}
}
\value{
List with solution data and status
}
\description{
Solve GLPK model using glpkAPI
}
