% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{export_import_log}
\alias{export_import_log}
\title{Export import log to CSV file}
\usage{
export_import_log(import_log, log_file)
}
\description{
Export import log to CSV file
}
\keyword{internal}
