% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{export_import_log_from_model}
\alias{export_import_log_from_model}
\title{Export import log to CSV}
\usage{
export_import_log_from_model(model, file)
}
\arguments{
\item{model}{Multimod model with import log}

\item{file}{Path to CSV file}
}
\description{
Export import log to CSV
}
