% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_r.R
\name{extract_args.ast_formula}
\alias{extract_args.ast_formula}
\title{Extract Arguments from ast_formula}
\usage{
\method{extract_args}{ast_formula}(x, args = character())
}
\arguments{
\item{x}{ast_formula object}

\item{args}{Character vector of accumulated argument names}
}
\value{
Character vector of argument names
}
\description{
Extract Arguments from ast_formula
}
