% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_r.R
\name{extract_args}
\alias{extract_args}
\title{Extract Arguments from AST}
\usage{
extract_args(x, args = character())
}
\arguments{
\item{x}{An AST node}

\item{args}{Character vector of accumulated argument names}
}
\value{
Character vector of unique argument names
}
\description{
Recursively extracts all symbol names (potential arguments) from an AST node.
This is used to automatically determine function arguments when converting
AST to R functions.
}
