% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gmpl.R
\name{extract_indexing}
\alias{extract_indexing}
\title{Extract Indexing Expression from GMPL Statement}
\usage{
extract_indexing(stmt)
}
\arguments{
\item{stmt}{Statement starting with indexing expression}
}
\value{
List with indexing structure and remaining statement
}
\description{
Parse {index_vars in sets: condition} pattern
}
\keyword{internal}
