% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{extract_mappings_from_equation}
\alias{extract_mappings_from_equation}
\title{Extract mapping names from equation AST}
\usage{
extract_mappings_from_equation(eq_obj)
}
\arguments{
\item{eq_obj}{Equation object}
}
\value{
Character vector of mapping names
}
\description{
Extract mapping names from equation AST
}
\keyword{internal}
