% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{extract_parameters_from_equation}
\alias{extract_parameters_from_equation}
\title{Extract parameter names from equation AST}
\usage{
extract_parameters_from_equation(eq_obj)
}
\arguments{
\item{eq_obj}{Equation object}
}
\value{
Character vector of parameter names
}
\description{
Extract parameter names from equation AST
}
\keyword{internal}
