% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_jump.R
\name{extract_parameters_from_expr}
\alias{extract_parameters_from_expr}
\title{Extract all parameter nodes from expression tree}
\usage{
extract_parameters_from_expr(node)
}
\description{
Extract all parameter nodes from expression tree
}
\keyword{internal}
