% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_parameters.R
\name{extract_set_members}
\alias{extract_set_members}
\title{Extract the member values of a set}
\usage{
extract_set_members(set_obj)
}
\description{
Extract the member values of a set
}
\keyword{internal}
