% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{extract_variables_from_equation}
\alias{extract_variables_from_equation}
\title{Extract variable names from equation AST}
\usage{
extract_variables_from_equation(eq_obj)
}
\arguments{
\item{eq_obj}{Equation object}
}
\value{
Character vector of variable names
}
\description{
Extract variable names from equation AST
}
\keyword{internal}
