% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{find_unmatched_elements}
\alias{find_unmatched_elements}
\title{Find unmatched elements between model and scenario}
\usage{
find_unmatched_elements(model, scenario, import_log)
}
\description{
Find unmatched elements between model and scenario
}
\keyword{internal}
