% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold.R
\name{fold_model_loop}
\alias{fold_model_loop}
\title{Internal fold loop with optional progress reporting}
\usage{
fold_model_loop(model, fold_spec, tolerance, verbose, p = NULL)
}
\description{
Internal fold loop with optional progress reporting
}
\keyword{internal}
