% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_gams.R
\name{format_gams_expression}
\alias{format_gams_expression}
\title{Format GAMS expression for readability}
\usage{
format_gams_expression(gams_expr, indent = 4, max_line_width = 80)
}
\arguments{
\item{gams_expr}{a character string representing a GAMS RHS expression}

\item{indent}{indentation per level (default: 4 spaces)}

\item{max_line_width}{approximate max line width before breaking}
}
\value{
character vector with formatted lines
}
\description{
Format GAMS expression for readability
}
