% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gmpl.R
\name{format_gmpl_expression}
\alias{format_gmpl_expression}
\title{Format GMPL expression with line breaks and indentation}
\usage{
format_gmpl_expression(gmpl_lines, indent = 2, ...)
}
\arguments{
\item{gmpl_lines}{Character vector of GMPL code lines}

\item{indent}{Number of spaces for indentation}

\item{...}{Additional arguments (not used)}
}
\value{
Formatted character vector
}
\description{
Format GMPL expression with line breaks and indentation
}
\keyword{internal}
