% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_latex.R
\name{format_latex_align}
\alias{format_latex_align}
\title{Format LaTeX equation for align* environment}
\usage{
format_latex_align(lhs, rel, rhs_string, max_line_length = 100)
}
\arguments{
\item{lhs}{Left-hand side of the equation (optional).}

\item{rel}{Relation string, such as '=', '<=', etc. Default is '='.}

\item{expr}{Full LaTeX expression string (optional).}

\item{rhs}{Right-hand side of the equation (optional).}

\item{max_width}{Approximate maximum character width before breaking line.}
}
\value{
A character vector of lines suitable for align* environment.
}
\description{
This function formats long LaTeX equations or expressions by inserting
line breaks and indentations to improve readability in the \code{align*} environment.
}
