% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_latex.R
\name{format_latex_equation_lines}
\alias{format_latex_equation_lines}
\title{Split a LaTeX equation body into multiple lines if it's too long}
\usage{
format_latex_equation_lines(lhs, rel, rhs, max_line_length = 120)
}
\arguments{
\item{max_line_length}{Approximate max number of characters per line.}

\item{body}{LaTeX string containing the full equation body (e.g., "lhs = rhs").}
}
\value{
A character vector of lines for align or align* block.
}
\description{
Split a LaTeX equation body into multiple lines if it's too long
}
