% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_latex.R
\name{format_set_elements}
\alias{format_set_elements}
\title{Format set elements for LaTeX display}
\usage{
format_set_elements(
  set_data,
  max_inline = 10,
  head_tail_threshold = 100,
  head_n = 3,
  tail_n = 3
)
}
\description{
Format set elements for LaTeX display
}
\keyword{internal}
