% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_aliases.R
\name{generate_index_aliases}
\alias{generate_index_aliases}
\title{Generate dummy variable names for a list of sets}
\usage{
generate_index_aliases(
  set_names,
  all_set_names = set_names,
  all_symbols = NULL
)
}
\arguments{
\item{set_names}{Character vector. Names of the sets}

\item{all_set_names}{Character vector. All set names in model (for alias detection)}

\item{all_symbols}{List. All symbols in model (from build_symbols_list)}
}
\value{
Named character vector. Dummy variable names with set names as names
}
\description{
Creates unique dummy variable names for a collection of sets
}
\examples{
generate_index_aliases(c("tech", "region", "comm", "commp", "year", "slice"))
}
