% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gmpl.R
\name{generate_variable_output}
\alias{generate_variable_output}
\title{Generate GMPL output code for a variable}
\usage{
generate_variable_output(var)
}
\arguments{
\item{var}{A variable object with name and dims}
}
\value{
Character vector of GMPL output code
}
\description{
Creates printf statements to export variable values to CSV file
}
\keyword{internal}
