% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{get_data_path}
\alias{get_data_path}
\title{Get data path for a parameter or mapping}
\usage{
get_data_path(obj, base_path = NULL)
}
\arguments{
\item{obj}{Parameter or mapping object}

\item{base_path}{Optional base path to prepend for relative paths}
}
\value{
Character path to data directory
}
\description{
Get data path for a parameter or mapping
}
