% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold.R, R/model_redundancy.R
\name{get_dim_names}
\alias{get_dim_names}
\title{Get dimension names from dims object
Get dimension names from dims object}
\usage{
get_dim_names(dims)

get_dim_names(dims)
}
\arguments{
\item{dims}{dims object or character vector}
}
\value{
Character vector of dimension names

Character vector of dimension names
}
\description{
Get dimension names from dims object
Get dimension names from dims object

Get dimension names from dims object (internal helper)
}
\keyword{internal}
