% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glpk_solver.R
\name{get_glpk_model_stats}
\alias{get_glpk_model_stats}
\title{Get GLPK model statistics without solving}
\usage{
get_glpk_model_stats(
  model_dir,
  mod_file = NULL,
  dat_file = NULL,
  verbose = TRUE
)
}
\arguments{
\item{model_dir}{Path to directory containing .mod and .dat files}

\item{mod_file}{Name of .mod file (default: searches for *.mod)}

\item{dat_file}{Name of .dat file (default: searches for *.dat)}

\item{verbose}{Logical, print statistics (default: TRUE)}
}
\value{
List with model statistics (n_rows, n_cols, n_nonzeros)
}
\description{
Get GLPK model statistics without solving
}
