% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{get_import_log}
\alias{get_import_log}
\title{Get data import log from model}
\usage{
get_import_log(model)
}
\arguments{
\item{model}{Multimod model with import log}
}
\value{
Import log list or NULL if not available
}
\description{
Get data import log from model
}
