% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_aliases.R
\name{get_index_alias}
\alias{get_index_alias}
\title{Get index alias for a set}
\usage{
get_index_alias(model, set_name)
}
\arguments{
\item{model}{A multimod model object}

\item{set_name}{Character. Name of the set}
}
\value{
Character. The index alias for the set
}
\description{
Retrieves the index alias for a set name from a multimod model.
If no alias exists, generates one on-the-fly.
}
\examples{
model <- add_index_aliases(model)
get_index_alias(model, "tech")  # "h"

}
