% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_latex.R
\name{get_latex_dims}
\alias{get_latex_dims}
\title{Get effective dimensions for LaTeX rendering}
\usage{
get_latex_dims(x, use_folded = TRUE)
}
\description{
Get effective dimensions for LaTeX rendering
}
\keyword{internal}
