% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{get_lazy_data}
\alias{get_lazy_data}
\title{Get lazy data with automatic loading}
\usage{
get_lazy_data(obj, base_path = NULL, collect = TRUE)
}
\arguments{
\item{obj}{Parameter or mapping object}

\item{base_path}{Base path for relative paths}

\item{collect}{Logical. Collect into memory?}
}
\value{
Data (data.table if collect=TRUE, Arrow dataset if FALSE)
}
\description{
Get lazy data with automatic loading
}
