% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_gmpl.R
\name{get_op_precedence}
\alias{get_op_precedence}
\title{Get operator precedence level (higher = tighter binding)}
\usage{
get_op_precedence(op)
}
\description{
Get operator precedence level (higher = tighter binding)
}
\keyword{internal}
