% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_parameters.R
\name{has_ast_formulas}
\alias{has_ast_formulas}
\title{Check if model has unevaluated AST formulas or defaults}
\usage{
has_ast_formulas(model)
}
\arguments{
\item{model}{Model object}
}
\value{
Logical
}
\description{
Check if model has unevaluated AST formulas or defaults
}
