% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_aliases.R
\name{has_index_aliases}
\alias{has_index_aliases}
\title{Check if model has index aliases}
\usage{
has_index_aliases(model)
}
\arguments{
\item{model}{A multimod model object}
}
\value{
Logical. TRUE if model has index_aliases
}
\description{
Check if model has index aliases
}
