% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_osemosys.R
\name{import_osemosys_data}
\alias{import_osemosys_data}
\title{Import OSeMOSYS data into multimod model}
\usage{
import_osemosys_data(model, dat_file, name_mapping = NULL, verbose = TRUE)
}
\arguments{
\item{model}{Multimod model (from read_gams)}

\item{dat_file}{Path to OSeMOSYS .dat file}

\item{name_mapping}{Optional named list mapping OSeMOSYS names to multimod names}

\item{verbose}{Logical. Print progress?}
}
\value{
Modified model with data attached
}
\description{
Import OSeMOSYS data into multimod model
}
\details{
Reads OSeMOSYS .dat file and attaches data to corresponding model elements.
For sets, populates set$data. For parameters, attaches parameter$data.

Name mapping example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{name_mapping <- list(
  TECHNOLOGY = "tech",
  REGION = "region",
  YEAR = "year"
)
}\if{html}{\out{</div>}}
}
