% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{is_in_memory}
\alias{is_in_memory}
\title{Data management utilities for multimod models}
\usage{
is_in_memory(obj)
}
\arguments{
\item{obj}{A multimod model or parameter/mapping object}
}
\value{
Logical. TRUE if data is in memory, FALSE if on disk
}
\description{
Functions to manage in-memory vs on-disk data storage for model parameters
and mappings, following the energyRt pattern.
Check if data is in memory
}
\details{
Checks hierarchy:
\enumerate{
\item Object's misc$inMemory if set
\item Model's inMemory default if object is a parameter/mapping
\item TRUE by default (conservative)
}
}
