% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{is_trimmed}
\alias{is_trimmed}
\title{Check if model or object is trimmed}
\usage{
is_trimmed(x)
}
\arguments{
\item{x}{Model object, or individual set/parameter/mapping/variable/equation}
}
\value{
Logical indicating if trimmed
}
\description{
Check if model or object is trimmed
}
\examples{
\dontrun{
is_trimmed(model)
is_trimmed(model$parameters$pDemand)
}

}
