% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{link_scenario_data_with_log}
\alias{link_scenario_data_with_log}
\title{Link scenario data with logging}
\usage{
link_scenario_data_with_log(model, scenario, inMemory = FALSE)
}
\arguments{
\item{model}{Multimod model}

\item{scenario}{energyRt scenario object}

\item{inMemory}{Logical}
}
\value{
List with model and logs
}
\description{
Link scenario data with logging
}
\keyword{internal}
