% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{load_all_mappings}
\alias{load_all_mappings}
\title{Load all mappings into memory}
\usage{
load_all_mappings(model)
}
\arguments{
\item{model}{Multimod model}
}
\value{
Modified model
}
\description{
Load all mappings into memory
}
