% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{load_all_parameters}
\alias{load_all_parameters}
\title{Load all parameters into memory}
\usage{
load_all_parameters(model)
}
\arguments{
\item{model}{Multimod model}
}
\value{
Modified model
}
\description{
Load all parameters into memory
}
