% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{load_arrow_data}
\alias{load_arrow_data}
\title{Load data from disk using Arrow}
\usage{
load_arrow_data(path, format = NULL, collect = TRUE)
}
\arguments{
\item{path}{Path to data directory or file}

\item{format}{Optional format specification}

\item{collect}{Logical. If TRUE, collect into memory (data.table)}
}
\value{
Arrow dataset or data.table
}
\description{
Load data from disk using Arrow
}
\keyword{internal}
