% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{load_mapping}
\alias{load_mapping}
\title{Load mapping data into memory}
\usage{
load_mapping(model, mapping_name)
}
\arguments{
\item{model}{Multimod model}

\item{mapping_name}{Name of mapping to load}
}
\value{
Modified model with mapping data loaded
}
\description{
Load mapping data into memory
}
