% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_osemosys.R
\name{load_osemosys_from_csv}
\alias{load_osemosys_from_csv}
\title{Load OSeMOSYS data from CSV directory}
\usage{
load_osemosys_from_csv(
  model,
  csv_dir,
  sets_subdir = "sets",
  params_subdir = "parameters",
  verbose = TRUE
)
}
\arguments{
\item{model}{Multimod model}

\item{csv_dir}{Directory containing CSV files}

\item{sets_subdir}{Subdirectory with set files}

\item{params_subdir}{Subdirectory with parameter files}

\item{verbose}{Logical}
}
\value{
Modified model with data loaded
}
\description{
Load OSeMOSYS data from CSV directory
}
