% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{load_parameter}
\alias{load_parameter}
\title{Load parameter data into memory}
\usage{
load_parameter(model, param_name)
}
\arguments{
\item{model}{Multimod model}

\item{param_name}{Name of parameter to load}
}
\value{
Modified model with parameter data loaded
}
\description{
Load parameter data into memory
}
