% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_ast}
\alias{map_ast}
\title{Recursively apply a function to all nodes of an AST}
\usage{
map_ast(node, func, include_class = TRUE)
}
\arguments{
\item{node}{An AST node or list of nodes.}

\item{func}{A function to apply to each node. It should return a scalar or structured value.}

\item{include_class}{Logical: if TRUE, include class name as part of result tree.}
}
\value{
A tree with the same structure but values replaced by func(node)
}
\description{
Recursively apply a function to all nodes of an AST
}
