% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{mark_empty_data}
\alias{mark_empty_data}
\title{Mark sets, parameters, and mappings with empty data}
\usage{
mark_empty_data(model, verbose = FALSE)
}
\arguments{
\item{model}{A multimod model object}

\item{verbose}{Logical. Print progress messages}
}
\value{
Model with \verb{$trimmed = TRUE} on empty elements
}
\description{
Mark sets, parameters, and mappings with empty data
}
\keyword{internal}
