% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{mark_empty_equations}
\alias{mark_empty_equations}
\title{Mark equations with empty domains or all variables trimmed}
\usage{
mark_empty_equations(model, verbose = FALSE)
}
\arguments{
\item{model}{A multimod model object}

\item{verbose}{Logical. Print progress messages}
}
\value{
Model with \verb{$trimmed = TRUE} on equations with empty domains
}
\description{
Mark equations with empty domains or all variables trimmed
}
\keyword{internal}
