% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{mark_empty_variables}
\alias{mark_empty_variables}
\title{Mark variables with empty domains}
\usage{
mark_empty_variables(model, verbose = FALSE)
}
\arguments{
\item{model}{A multimod model object}

\item{verbose}{Logical. Print progress messages}
}
\value{
Model with \verb{$trimmed = TRUE} on variables with empty domains
}
\description{
Checks if any dimension of a variable references an empty or trimmed set.
}
\keyword{internal}
