% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gmpl.R
\name{matching_paren}
\alias{matching_paren}
\title{Find matching closing parenthesis}
\usage{
matching_paren(expr)
}
\arguments{
\item{expr}{Character string starting with opening paren}
}
\value{
Position of matching closing paren, or NA if not found
}
\description{
Find matching closing parenthesis
}
\keyword{internal}
