% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_parameters.R
\name{materialize_parameter_values}
\alias{materialize_parameter_values}
\title{Convert parameter data frames into arrays or vectors for indexing}
\usage{
materialize_parameter_values(param, data = NULL, dim_members = list())
}
\description{
Convert parameter data frames into arrays or vectors for indexing
}
\keyword{internal}
