% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimod-classes.R
\name{new_mapping}
\alias{new_mapping}
\title{Create a multimod mapping object}
\usage{
new_mapping(
  name,
  desc = NULL,
  subset_of = NULL,
  dims = NULL,
  active_dims = NULL,
  data = NULL,
  auto_fold = FALSE,
  symbols = NULL
)
}
\arguments{
\item{name}{character, name of the mapping}

\item{desc}{character, description of the mapping}

\item{subset_of}{character, name of the parent mapping}

\item{dims}{character vector, names of the declared dimensions}

\item{data}{data frame, data for the mapping}
}
\value{
An object of class \code{mapping}
}
\description{
Create a multimod mapping object
}
