% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimod-classes.R
\name{new_model}
\alias{new_model}
\title{Create a multimod model object}
\usage{
new_model(
  name = NULL,
  desc = NULL,
  sets = list(),
  aliases = list(),
  mappings = list(),
  parameters = list(),
  variables = list(),
  equations = list(),
  inMemory = TRUE,
  base_path = NULL,
  metadata = list(),
  ...
)
}
\arguments{
\item{desc}{model description (character)}

\item{sets}{named list of sets}

\item{mappings}{named list of mappings}

\item{parameters}{named list of parameter objects}

\item{variables}{named list of variable objects}

\item{equations}{named list of equation objects}

\item{metadata}{Named list with auxiliary information (e.g., language, source_file, data_source).}
}
\value{
A model object
}
\description{
Create a multimod model object
}
