% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimod.R
\name{new_multimod_eqn}
\alias{new_multimod_eqn}
\title{Create a multimod equation object}
\usage{
new_multimod_eqn(
  name,
  declared_dims,
  lhs,
  rhs,
  sense = "==",
  domain = NULL,
  desc = NULL
)
}
\arguments{
\item{desc}{}
}
\description{
Create a multimod equation object
}
