% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimod.R
\name{new_multimod_model}
\alias{new_multimod_model}
\title{Create a multimod model object}
\usage{
new_multimod_model(
  sets = list(),
  mappings = list(),
  parameters = list(),
  variables = list(),
  equations = list(),
  desc = NULL
)
}
\arguments{
\item{sets}{named list of sets}

\item{mappings}{named list of mappings}

\item{parameters}{named list of multimod_param objects}

\item{variables}{named list of multimod_var objects}

\item{equations}{named list of multimod_eqn objects}

\item{desc}{}
}
\description{
Create a multimod model object
}
