% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimod.R
\name{new_multimod_param}
\alias{new_multimod_param}
\title{Create a multimod parameter}
\usage{
new_multimod_param(
  name,
  declared_dims,
  data,
  active_dims = NULL,
  auto_fold = FALSE
)
}
\arguments{
\item{name}{character, name of the parameter}

\item{declared_dims}{character vector, names of the declared dimensions}

\item{data}{data frame, data for the parameter}

\item{active_dims}{character vector, names of the active dimensions}

\item{auto_fold}{logical, whether to automatically fold dimensions}
}
\value{
a multimod_param object
}
\description{
Create a multimod parameter
}
