% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimod-classes.R
\name{new_multimod_set}
\alias{new_multimod_set}
\alias{new_set}
\title{Create a new multimod_set object}
\usage{
new_multimod_set(name, desc = NULL, subset_of = NULL, data = NULL)
}
\arguments{
\item{name}{Character. Name of the set.}

\item{desc}{Character. Description of the set (optional).}

\item{subset_of}{Character vector. Name of the parent set (optional).}

\item{data}{Optional values (e.g., character vector or data.frame).}
}
\value{
An object of class \code{multimod_set}
}
\description{
Create a new multimod_set object
}
