% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimod-classes.R
\name{new_multimod_variable}
\alias{new_multimod_variable}
\alias{new_variable}
\title{Create a multimod variable}
\usage{
new_multimod_variable(
  name,
  desc = NULL,
  dims,
  active_dims = NULL,
  domain = NULL,
  data = NULL,
  auto_fold = FALSE
)
}
\arguments{
\item{name}{character, name of the variable}

\item{dims}{character vector, names of the declared dimensions}

\item{active_dims}{character vector, names of the active dimensions}

\item{domain}{character, domain of the variable (e.g., "continuous", "integer", "binary")}

\item{data}{data frame, data for the variable}

\item{auto_fold}{logical, whether to automatically fold dimensions}
}
\value{
a multimod_variable object
}
\description{
Create a multimod variable
}
