% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimod-classes.R
\name{new_parameter}
\alias{new_parameter}
\title{Create a multimod parameter}
\usage{
new_parameter(
  name,
  desc = NULL,
  dims,
  active_dims = NULL,
  data = NULL,
  defVal = NULL,
  defInt = NULL,
  symbolic = FALSE,
  formula = NULL,
  comment = NULL,
  auto_fold = FALSE,
  symbols = NULL
)
}
\arguments{
\item{name}{character, name of the parameter}

\item{dims}{character vector, names of the declared dimensions}

\item{active_dims}{character vector, names of the active dimensions}

\item{data}{data frame, data for the parameter}

\item{comment}{character, comment line from source file (e.g., *@ domain hint)}

\item{auto_fold}{logical, whether to automatically fold dimensions}
}
\value{
a parameter object
}
\description{
Create a multimod parameter
}
