% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast-classes.R
\name{node_type}
\alias{node_type}
\title{Get the type of an AST or multimod node}
\usage{
node_type(x)
}
\arguments{
\item{x}{An \code{ast} or \code{multimod} object. In case of other classes, it will
return \code{NULL}.}
}
\value{
A character string representing the type of the node
(e.g., \code{"expression"}, \code{"variable"}, \code{"parameter"}, \code{"sum"}, etc.).
}
\description{
This function retrieves the class of an AST node or multimod object that
represents a specific type of node in the abstract syntax tree.
}
