% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_parameters.R
\name{normalize_dim_names}
\alias{normalize_dim_names}
\title{Normalize dimension specifications to a character vector of set names}
\usage{
normalize_dim_names(dims)
}
\description{
Normalize dimension specifications to a character vector of set names
}
\keyword{internal}
