% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_parameters.R
\name{normalize_index_aliases}
\alias{normalize_index_aliases}
\title{Normalize provided dimension aliases to align with dim names}
\usage{
normalize_index_aliases(aliases, dim_names)
}
\description{
Normalize provided dimension aliases to align with dim names
}
\keyword{internal}
