% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_osemosys.R
\name{osemosys_dat_to_model}
\alias{osemosys_dat_to_model}
\title{Convert OSeMOSYS .dat to CSV and import into model}
\usage{
osemosys_dat_to_model(
  model,
  dat_file,
  csv_dir = NULL,
  name_mapping = NULL,
  verbose = TRUE
)
}
\arguments{
\item{model}{Multimod model}

\item{dat_file}{Path to OSeMOSYS .dat file}

\item{csv_dir}{Directory to save CSV files (optional)}

\item{name_mapping}{Optional name mapping}

\item{verbose}{Logical}
}
\value{
Modified model with data attached
}
\description{
Convert OSeMOSYS .dat to CSV and import into model
}
\details{
Convenience function that:
\enumerate{
\item Reads OSeMOSYS .dat file
\item Optionally exports to CSV
\item Imports data into model
}
}
