% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_parameters.R
\name{parameter_result_to_dataframe}
\alias{parameter_result_to_dataframe}
\title{Convert evaluated parameter arrays into long-form data frames}
\usage{
parameter_result_to_dataframe(result, dims)
}
\description{
Convert evaluated parameter arrays into long-form data frames
}
\keyword{internal}
