% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gmpl.R
\name{parse_bracket_format}
\alias{parse_bracket_format}
\title{Parse bracket format for parameter values}
\usage{
parse_bracket_format(name, text, default_val)
}
\description{
Parses GMPL bracket format like \verb{[idx1,idx2] value [idx1,idx2] value}.
}
\keyword{internal}
