% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gmpl.R
\name{parse_function_call}
\alias{parse_function_call}
\title{Parse function call}
\usage{
parse_function_call(expr, symbols, depth, max_depth)
}
\arguments{
\item{expr}{Character string}

\item{symbols}{Symbol table}

\item{depth}{Current depth}

\item{max_depth}{Maximum depth}
}
\value{
AST node or NULL
}
\description{
Parse function call
}
\keyword{internal}
