% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gmpl.R
\name{parse_gmpl_declarations}
\alias{parse_gmpl_declarations}
\title{Parse GMPL Declarations}
\usage{
parse_gmpl_declarations(text, model, verbose = FALSE)
}
\arguments{
\item{text}{Preprocessed GMPL text}

\item{model}{Model object to populate}

\item{verbose}{Logical, print progress}
}
\value{
Updated model object
}
\description{
Main parser that identifies and processes all declarations
}
\keyword{internal}
