% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gmpl.R
\name{parse_gmpl_equation}
\alias{parse_gmpl_equation}
\title{Parse GMPL Equation Declaration}
\usage{
parse_gmpl_equation(stmt, model, verbose = FALSE)
}
\arguments{
\item{stmt}{Equation declaration statement (s.t. constraint)}

\item{model}{Model object}

\item{verbose}{Logical, print progress}
}
\value{
Updated model object
}
\description{
Parse GMPL Equation Declaration
}
\keyword{internal}
