% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gmpl.R
\name{parse_gmpl_objective}
\alias{parse_gmpl_objective}
\title{Parse GMPL Objective Declaration}
\usage{
parse_gmpl_objective(stmt, model, verbose = FALSE)
}
\arguments{
\item{stmt}{Objective declaration statement}

\item{model}{Model object}

\item{verbose}{Logical, print progress}
}
\value{
Updated model object
}
\description{
Parse GMPL Objective Declaration
}
\keyword{internal}
