% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gmpl.R
\name{parse_gmpl_param}
\alias{parse_gmpl_param}
\title{Parse GMPL Parameter Declaration}
\usage{
parse_gmpl_param(stmt, model, verbose = FALSE)
}
\arguments{
\item{stmt}{Parameter declaration statement}

\item{model}{Model object}

\item{verbose}{Logical, print progress}
}
\value{
Updated model object
}
\description{
Parse GMPL Parameter Declaration
}
\keyword{internal}
