% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gmpl.R
\name{parse_gmpl_set}
\alias{parse_gmpl_set}
\title{Parse GMPL Set Declaration}
\usage{
parse_gmpl_set(stmt, model, verbose = FALSE)
}
\arguments{
\item{stmt}{Set declaration statement}

\item{model}{Model object}

\item{verbose}{Logical, print progress}
}
\value{
Updated model object
}
\description{
Parse GMPL Set Declaration
}
\keyword{internal}
