% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gmpl.R
\name{parse_gmpl_var}
\alias{parse_gmpl_var}
\title{Parse GMPL Variable Declaration}
\usage{
parse_gmpl_var(stmt, model, verbose = FALSE)
}
\arguments{
\item{stmt}{Variable declaration statement}

\item{model}{Model object}

\item{verbose}{Logical, print progress}
}
\value{
Updated model object
}
\description{
Parse GMPL Variable Declaration
}
\keyword{internal}
