% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gmpl.R
\name{parse_index_list}
\alias{parse_index_list}
\title{Parse Index List}
\usage{
parse_index_list(str)
}
\arguments{
\item{str}{Index list string}
}
\value{
List of index specifications
}
\description{
Parse comma-separated index specifications like "i in SET, j in SET2"
}
\keyword{internal}
