% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gmpl.R
\name{parse_list_format}
\alias{parse_list_format}
\title{Parse list format: idx value idx value}
\usage{
parse_list_format(name, text, default_val)
}
\description{
Parse list format: idx value idx value
}
\keyword{internal}
