% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_osemosys.R
\name{parse_osemosys_parameter}
\alias{parse_osemosys_parameter}
\title{Parse OSeMOSYS parameter declaration}
\usage{
parse_osemosys_parameter(decl)
}
\arguments{
\item{decl}{Parameter declaration string}
}
\value{
List with name and data frame, or NULL if parsing fails
}
\description{
Parse OSeMOSYS parameter declaration
}
\keyword{internal}
