% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_osemosys.R
\name{parse_osemosys_set}
\alias{parse_osemosys_set}
\title{Parse OSeMOSYS set declaration}
\usage{
parse_osemosys_set(decl)
}
\arguments{
\item{decl}{Set declaration string}
}
\value{
List with name and data, or NULL if parsing fails
}
\description{
Parse OSeMOSYS set declaration
}
\keyword{internal}
