% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gmpl.R
\name{parse_param_declaration}
\alias{parse_param_declaration}
\title{Parse parameter declaration from GMPL data file}
\usage{
parse_param_declaration(lines, start_line)
}
\arguments{
\item{lines}{All lines from the file}

\item{start_line}{Line number where parameter declaration starts}
}
\value{
List with name, type, dims, default, data, comment, next_line
}
\description{
Parse parameter declaration from GMPL data file
}
\keyword{internal}
