% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gmpl.R
\name{parse_set_declaration}
\alias{parse_set_declaration}
\title{Parse set declaration from GMPL data file}
\usage{
parse_set_declaration(lines, start_line)
}
\arguments{
\item{lines}{All lines from the file}

\item{start_line}{Line number where set declaration starts}
}
\value{
List with name, type, dims, data, next_line
}
\description{
Parse set declaration from GMPL data file
}
\keyword{internal}
