% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gmpl.R
\name{parse_simple_table_format}
\alias{parse_simple_table_format}
\title{Parse simple table format: : cols := rows values}
\usage{
parse_simple_table_format(name, header, text, default_val)
}
\description{
Parse simple table format: : cols := rows values
}
\keyword{internal}
