% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.multimod_ast}
\alias{plot.multimod_ast}
\title{Plot a multimod AST using visNetwork}
\usage{
\method{plot}{multimod_ast}(x, ...)
}
\arguments{
\item{x}{A multimod_ast object}

\item{...}{Additional arguments (currently unused)}
}
\value{
A visNetwork plot
}
\description{
Plot a multimod AST using visNetwork
}
