% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_energyrt.R
\name{populate_defvals_from_energyrt}
\alias{populate_defvals_from_energyrt}
\title{Update model parameters with default values from energyRt}
\usage{
populate_defvals_from_energyrt(model)
}
\arguments{
\item{model}{multimod model object}
}
\value{
Updated model with defVal populated
}
\description{
Update model parameters with default values from energyRt
}
