% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{populate_sets_from_scenario}
\alias{populate_sets_from_scenario}
\title{Populate sets from energyRt scenario}
\usage{
populate_sets_from_scenario(model, scenario)
}
\arguments{
\item{model}{Multimod model}

\item{scenario}{energyRt scenario object}
}
\value{
Modified model with set data populated
}
\description{
Populate sets from energyRt scenario
}
\details{
Extracts unique members for each set by scanning through all parameters
and mappings in the energyRt scenario. For each dimension that matches
a set name, collects unique values.
}
