% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{populate_sets_with_log}
\alias{populate_sets_with_log}
\title{Populate sets and track data sources}
\usage{
populate_sets_with_log(model, scenario)
}
\arguments{
\item{model}{Multimod model}

\item{scenario}{energyRt scenario object}
}
\value{
List with model and log
}
\description{
Populate sets and track data sources
}
\keyword{internal}
