% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_aliases.R
\docType{data}
\name{preferred_dummy_names}
\alias{preferred_dummy_names}
\title{Preferred short names for common sets to avoid conflicts}
\format{
An object of class \code{list} of length 13.
}
\usage{
preferred_dummy_names
}
\description{
Preferred short names for common sets to avoid conflicts
}
\keyword{internal}
