% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gmpl.R
\name{preprocess_gmpl}
\alias{preprocess_gmpl}
\title{Preprocess GMPL Text}
\usage{
preprocess_gmpl(lines)
}
\arguments{
\item{lines}{Character vector of file lines}
}
\value{
Preprocessed lines
}
\description{
Remove comments and normalize whitespace
}
\keyword{internal}
