% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_r.R
\name{print.ast_function}
\alias{print.ast_function}
\title{Print method for functions created from AST}
\usage{
\method{print}{ast_function}(x, ...)
}
\arguments{
\item{x}{A function created by as_rfunction}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for functions created from AST
}
