% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.ast}
\alias{print.func}
\title{Print methods for AST classes}
\usage{
\method{print}{ast}(x, ...)

\method{print}{func}(x, ...)
}
\arguments{
\item{x}{Object to print}

\item{...}{Additional arguments passed to print methods}
}
\description{
S3 print methods for displaying AST and multimod objects
}
