% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_solutions.R
\name{print.variable_comparison}
\alias{print.variable_comparison}
\title{Print method for variable comparison results}
\usage{
\method{print}{variable_comparison}(x, ...)
}
\arguments{
\item{x}{A variable_comparison object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for variable comparison results
}
