% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_osemosys.R
\name{read_osemosys_dat}
\alias{read_osemosys_dat}
\title{Read OSeMOSYS data files}
\usage{
read_osemosys_dat(dat_file, verbose = TRUE)
}
\arguments{
\item{dat_file}{Path to .dat file}

\item{verbose}{Logical. Print progress messages?}
}
\value{
List with sets and parameters as data frames
}
\description{
Functions to import OSeMOSYS data in various formats (Pyomo .dat, AMPL .dat, etc.)
and convert to multimod format.
Parse OSeMOSYS .dat file (Pyomo/AMPL format)
}
\details{
Parses OSeMOSYS data files in Pyomo/AMPL format. Handles:
\itemize{
\item Set declarations: \verb{set SETNAME := value1 value2 ... ;}
\item Parameters with dimensions: \verb{param ParamName := key1 key2 value ;}
\item Multi-line continuations
\item Comments starting with #
}
}
