% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remap_ast_elements}
\alias{remap_ast_elements}
\title{Replace AST elements with "ast_where" nodes}
\usage{
remap_ast_elements(
  obj,
  ast_type = list(when = "condition"),
  name_prefix = "m",
  n = 0L,
  latex_max = 10,
  ...
)
}
\arguments{
\item{obj}{An object of class \code{ast}, \code{multimod}, or a list.}

\item{ast_type}{A character vector of AST types to be replaced
(e.g., "when", "mapping").}

\item{name_prefix}{A character string prefix for the new names of the
replaced elements. Defaults to "m".}
}
\value{
Modified AST object with remapped elements
}
\description{
This function replaces elements of a given AST object with "ast_where" nodes.
It is useful for long expressions to improve readability of latex output and
network trees. The "ast_where" stores the original expression in a \verb{$content}
slot, and ignored by the parsers to GAMS, Julia, and other languages.
}
