% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{remove_unused_variables}
\alias{remove_unused_variables}
\title{Remove variables not appearing in any non-trimmed equation}
\usage{
remove_unused_variables(model, verbose = FALSE)
}
\arguments{
\item{model}{A multimod model object}

\item{verbose}{Logical. Print progress messages}
}
\value{
Model with unused variables trimmed
}
\description{
Final pass to mark variables that don't appear in any active equation.
}
\keyword{internal}
