% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_latex.R
\name{render_extrema_index}
\alias{render_extrema_index}
\title{Render LaTeX indices for extrema constructs}
\usage{
render_extrema_index(index_node, brackets = NULL, ...)
}
\description{
Render LaTeX indices for extrema constructs
}
\keyword{internal}
