% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_latex.R
\name{replace_mapping_placeholders}
\alias{replace_mapping_placeholders}
\title{Replace mapping placeholders like \texttt{m1} with styled math symbols}
\usage{
replace_mapping_placeholders(
  latex_str,
  mathcal = FALSE,
  symbol = ifelse(mathcal, "M", "psi")
)
}
\arguments{
\item{latex_str}{A character vector of LaTeX lines.}

\item{style}{One of "mathcal", "psi", "phi", or "custom".}

\item{base}{Optional base symbol for style = "custom" (e.g., "M").}
}
\value{
Updated LaTeX character vector.
}
\description{
Replace mapping placeholders like \texttt{m1} with styled math symbols
}
