% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resolve_alias}
\alias{resolve_alias}
\title{Resolve a dimension name to its alias}
\usage{
resolve_alias(
  name,
  base = base_aliases,
  extended = extended_aliases,
  default = NULL
)
}
\arguments{
\item{name}{A character string or vector of dimension names.}

\item{base}{Named list of base aliases.}

\item{extended}{Named list of extended aliases.}

\item{default}{If no alias is found, return \code{name}, \code{NA}, or \code{"?"}.}
}
\value{
A character vector of resolved aliases.
}
\description{
Resolve a dimension name to its alias
}
