% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_parameters.R
\name{resolve_index_vars}
\alias{resolve_index_vars}
\title{Resolve iterator variable names for dimensions}
\usage{
resolve_index_vars(dims_index_aliases, dim_names)
}
\description{
Resolve iterator variable names for dimensions
}
\keyword{internal}
