% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_parameters.R
\name{restore_ast}
\alias{restore_ast}
\title{Restore AST from preserved copies}
\usage{
restore_ast(model)
}
\arguments{
\item{model}{Model with evaluated parameters}
}
\value{
Model with AST restored
}
\description{
Restore formula and defVal AST from param$misc storage
}
