% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{save_all_mappings}
\alias{save_all_mappings}
\title{Save all mappings to disk}
\usage{
save_all_mappings(model, format = "parquet", keep_in_memory = FALSE)
}
\arguments{
\item{model}{Multimod model}

\item{format}{Data format}

\item{keep_in_memory}{Keep data in memory after saving?}
}
\value{
Modified model
}
\description{
Save all mappings to disk
}
