% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{save_arrow_data}
\alias{save_arrow_data}
\title{Save data to disk using Arrow}
\usage{
save_arrow_data(data, path, format = "parquet", overwrite = FALSE)
}
\arguments{
\item{data}{Data frame or data.table to save}

\item{path}{Path to save to}

\item{format}{Format: "parquet", "feather", "ipc", "csv"}

\item{overwrite}{Logical. Overwrite existing data?}
}
\description{
Save data to disk using Arrow
}
\keyword{internal}
