% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{save_mapping}
\alias{save_mapping}
\title{Save mapping data to disk and free memory}
\usage{
save_mapping(model, mapping_name, format = "parquet", keep_in_memory = FALSE)
}
\arguments{
\item{model}{Multimod model}

\item{mapping_name}{Name of mapping to save}

\item{format}{Data format: "parquet", "feather", "ipc", "csv"}

\item{keep_in_memory}{Keep data in memory after saving?}
}
\value{
Modified model
}
\description{
Save mapping data to disk and free memory
}
