% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{save_parameter}
\alias{save_parameter}
\title{Save parameter data to disk and free memory}
\usage{
save_parameter(model, param_name, format = "parquet", keep_in_memory = FALSE)
}
\arguments{
\item{model}{Multimod model}

\item{param_name}{Name of parameter to save}

\item{format}{Data format: "parquet", "feather", "ipc", "csv"}

\item{keep_in_memory}{Keep data in memory after saving?}
}
\value{
Modified model
}
\description{
Save parameter data to disk and free memory
}
