% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{set_data_path}
\alias{set_data_path}
\title{Set data path for a parameter or mapping}
\usage{
set_data_path(obj, path)
}
\arguments{
\item{obj}{Parameter or mapping object}

\item{path}{Character path (relative or absolute)}
}
\value{
Modified object
}
\description{
Set data path for a parameter or mapping
}
