% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{show_bounds_mapping}
\alias{show_bounds_mapping}
\title{Show bounds parameter mapping}
\usage{
show_bounds_mapping(model)
}
\arguments{
\item{model}{Multimod model with import log}
}
\value{
Data frame showing bounds parameter splits
}
\description{
Show bounds parameter mapping
}
