% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{show_unmatched}
\alias{show_unmatched}
\title{Show unmatched elements}
\usage{
show_unmatched(model)
}
\arguments{
\item{model}{Multimod model with import log}
}
\description{
Show unmatched elements
}
