% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_latex.R
\name{split_at_top_level_operators}
\alias{split_at_top_level_operators}
\title{Split LaTeX math string at top-level operators}
\usage{
split_at_top_level_operators(
  latex_str,
  operators = c("+", "-", "\\\\cdot", "\\\\div", "="),
  indent_str = character(0),
  max_len = 80
)
}
\arguments{
\item{latex_str}{A LaTeX math string}

\item{operators}{Vector of operators to split at}
}
\value{
A character vector of expression chunks
}
\description{
This function splits a LaTeX math string at given types of top-level
operators and the maximum length of each chunk. The splitting is
done recursively with respect to the operator order.
}
