% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gmpl.R
\name{split_statements}
\alias{split_statements}
\title{Split Text into Statements}
\usage{
split_statements(text)
}
\arguments{
\item{text}{GMPL text}
}
\value{
Character vector of statements
}
\description{
Split GMPL text by semicolons, respecting nested braces
}
\keyword{internal}
