% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{split_top_level_operators}
\alias{split_top_level_operators}
\title{Split LaTeX math string at top-level operators}
\usage{
split_top_level_operators(
  latex_str,
  operators = c("+", "-", "\\\\\\\\cdot", "\\\\\\\\div", "=")
)
}
\arguments{
\item{latex_str}{A LaTeX math string}

\item{operators}{Vector of operators to split at}
}
\value{
A character vector of expression chunks, including the operators
}
\description{
Split LaTeX math string at top-level operators
}
