% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_storage.R
\name{update_mapping}
\alias{update_mapping}
\title{Update mapping data}
\usage{
update_mapping(model, name, data, auto_save = FALSE)
}
\arguments{
\item{model}{A multimod model object}

\item{name}{Character. Mapping name}

\item{data}{Data.frame with mapping tuples}

\item{auto_save}{Logical. If TRUE and model is on-disk, immediately save to disk}
}
\value{
Modified model object
}
\description{
Low-level function to update mapping data in a model.
}
\examples{
\dontrun{
# Update mapping data
model <- update_mapping(model, "mTechRegion", new_mapping_data)
}
}
