% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_character.R
\name{wrap_brackets}
\alias{wrap_brackets}
\title{Wrap text with brackets}
\usage{
wrap_brackets(x, brackets = NULL, ...)
}
\arguments{
\item{x}{Character string to wrap}

\item{brackets}{Bracket style to use}

\item{...}{Additional arguments}
}
\value{
Character string wrapped in brackets
}
\description{
Wrap text with brackets
}
