% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gams.R
\name{write_gams.multimod}
\alias{write_gams.multimod}
\title{Write a full GAMS model file from a multimod object}
\usage{
write_gams.multimod(model, file = NULL, format_expr = TRUE, ...)
}
\arguments{
\item{model}{A \code{multimod} model object}

\item{file}{Output file path (optional)}

\item{format_expr}{logical; whether to format expressions with line breaks}

\item{...}{Additional arguments passed to formatting functions}
}
\value{
Character vector or writes file if \code{file} is given
}
\description{
Write a full GAMS model file from a multimod object
}
