% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_jump.R
\name{write_jump_solution_code}
\alias{write_jump_solution_code}
\title{Write JuMP solution values to disk}
\usage{
write_jump_solution_code()
}
\arguments{
\item{model}{JuMP model object (from Julia)}

\item{output_dir}{Directory to save solutions}

\item{format}{Data format ("arrow" or "csv")}

\item{run_name}{Name for this solution run (default: timestamp)}
}
\description{
Helper function to save solution values from Julia model.
Called from within Julia code after solving.
}
\details{
This is typically called from within the generated model.jl file.
It saves variable values in the same directory structure as input data.
}
\examples{
\dontrun{
# In Julia (model.jl):
# After solve...
using Dates
run_name = Dates.format(now(), "yyyy-mm-dd_HHMMSS")
save_jump_solution(model, "../../solutions/" * run_name, DATA_FORMAT)
}

}
