% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_latex.R
\name{write_latex}
\alias{write_latex}
\title{Write LaTeX representation of an equation or model}
\usage{
write_latex(x, file, ...)
}
\arguments{
\item{x}{An object of class \code{equation} or \code{model}.}

\item{file}{Path to the output LaTeX file. If \code{NULL}, returns the LaTeX code as a character string.}

\item{...}{Additional arguments passed to the specific method.}
}
\description{
Write LaTeX representation of an equation or model
}
