% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gmpl.R
\name{write_mapping_data}
\alias{write_mapping_data}
\title{Write mapping data (tuples)}
\usage{
write_mapping_data(mapping_obj, model = NULL)
}
\arguments{
\item{mapping_obj}{Mapping object}

\item{model}{Model object (for lazy loading)}
}
\value{
Character vector of GMPL data statements
}
\description{
Write mapping data (tuples)
}
\keyword{internal}
