% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gmpl.R
\name{write_set_data}
\alias{write_set_data}
\title{Write set data (members)}
\usage{
write_set_data(set_obj, model = NULL)
}
\arguments{
\item{set_obj}{Set object}

\item{model}{Model object (for base_path if needed)}
}
\value{
Character vector of GMPL data statements
}
\description{
Write set data (members)
}
\keyword{internal}
