% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{alias_ast_names}
\alias{alias_ast_names}
\title{Apply name aliases to AST nodes}
\usage{
alias_ast_names(ast, alias_map, classes = NULL, ...)
}
\arguments{
\item{ast}{An AST object or list.}

\item{alias_map}{A named list or character vector. Keys are aliases, values are original names.}

\item{classes}{Character vector of AST classes to filter (optional).}
}
\value{
Modified AST with aliases applied.
}
\description{
Apply name aliases to AST nodes
}
