% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{apply_bounds_to_model}
\alias{apply_bounds_to_model}
\title{Attach bounds data to model parameters}
\usage{
apply_bounds_to_model(model, scenario_param, base_name)
}
\arguments{
\item{model}{Multimod model}

\item{scenario_param}{energyRt bounds parameter}

\item{base_name}{Base parameter name (without Lo/Up suffix)}
}
\value{
List with model and diagnostic info
}
\description{
Attach bounds data to model parameters
}
\keyword{internal}
