% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast-classes.R
\name{ast_parameter}
\alias{ast_parameter}
\title{Create a parameter AST node}
\usage{
ast_parameter(name, dims = ast_dims())
}
\arguments{
\item{name}{Name of the parameter.}

\item{dims}{Optional vector of dimension \code{ast_dims} objects.}
}
\value{
An \code{ast_parameter} S3 object (subclass of \code{multimod_ast}).
}
\description{
Create a parameter AST node
}
