% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast-classes.R
\name{ast_prod}
\alias{ast_prod}
\title{Create a product AST node}
\usage{
ast_prod(index, value)
}
\arguments{
\item{index}{Character. The index variable (e.g., \code{"t"}).}

\item{value}{An AST node representing the expression to be summed.}
}
\value{
An object of class \code{ast} and \code{prod}.
}
\description{
Constructs an abstract syntax tree (AST) node representing a product over an
index. The index is typically a \code{set}, \code{dims} or \code{when} object if
filtering is applied to the index.
}
