% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast-classes.R
\name{ast_sum}
\alias{ast_sum}
\title{Create a summation AST node}
\usage{
ast_sum(index = ast_dims(), value)
}
\arguments{
\item{index}{Character. The index variable (e.g., \code{"t"}).}

\item{value}{An AST node representing the expression to be summed.}
}
\value{
An object of class \code{ast} and \code{sum}.
}
\description{
Constructs an abstract syntax tree (AST) node representing a summation over an
index. The index is typically a \code{set}, \code{dims} or \code{when} object if
filtering is applied to the index.
}
