% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast-classes.R
\name{ast_sum}
\alias{ast_sum}
\title{Create a summation AST node}
\usage{
ast_sum(index = ast_dims(), domain = NULL, value)
}
\arguments{
\item{index}{Character. The index variable (e.g., \code{"t"}).}

\item{domain}{An optional AST node (class \code{ast}) representing a domain condition.
This can be a mapping, a logical condition (e.g., from a \code{$}-filter), or a parameter.
Use \code{NULL} if there is no restriction.}

\item{value}{An AST node representing the expression to be summed.}
}
\value{
An object of class \code{ast} and \code{sum}.
}
\description{
Constructs an abstract syntax tree (AST) node representing a summation over an index,
optionally restricted by a domain condition (such as a mapping, logical condition, or parameter).
}
