% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast-classes.R
\name{ast_variable}
\alias{ast_variable}
\title{Create a variable AST node}
\usage{
ast_variable(name, dims = ast_dims(), vtype = NULL, bounds = NULL, ...)
}
\arguments{
\item{name}{Name of the variable.}

\item{dims}{Optional vector of dimension \code{dims} objects.}
}
\value{
An \code{variable} S3 object (subclass of \code{ast}).
}
\description{
Create a variable AST node
}
