% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast-classes.R
\name{ast_where}
\alias{ast_where}
\title{Create a "where" AST node}
\usage{
ast_where(name, content, hash = node_hash(content), ...)
}
\arguments{
\item{name}{A character string representing the name of the reference,
matching the name of the symbol replacing AST node or a branch of the AST.}

\item{...}{}
}
\description{
Constructs a "where" node representing a reference to a specific
location in the abstract syntax tree (AST).
}
