% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_energyrt.R
\name{convert_energyrt_parameter}
\alias{convert_energyrt_parameter}
\title{Convert energyRt parameter to multimod format}
\usage{
convert_energyrt_parameter(
  ert_param,
  orig_param = NULL,
  inMemory = FALSE,
  scenario = NULL
)
}
\arguments{
\item{ert_param}{An energyRt parameter object}

\item{orig_param}{Original multimod parameter structure (optional)}

\item{inMemory}{Logical. Embed data in-memory?}
}
\value{
Updated parameter with data reference
}
\description{
Convert energyRt parameter to multimod format
}
\keyword{internal}
