% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimod-package.R
\docType{package}
\name{multimod-package}
\alias{multimod}
\alias{multimod-package}
\title{multimod: Define and Render Mathematical Models Across Multiple Optimization Languages.}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Multimod provides a domain-specific language (DSL) for defining algebraic equations and constraints in R, and rendering them to multiple mathematical programming languages, including GAMS, AMPL, Pyomo, and JuMP. It supports structured storage of equations, parameters, sets, and mappings, and aims to simplify the development of complex optimization models across platforms.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://optimal2050.github.io/multimod}
}

}
\author{
\strong{Maintainer}: Oleg Lugovoy \email{olugovoy@optimalsolution.dev} (\href{https://orcid.org/0000-0001-5546-9875}{ORCID})

}
\keyword{internal}
