% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimod-classes.R
\name{new_multimod_model}
\alias{new_multimod_model}
\title{Create a multimod model object}
\usage{
new_multimod_model(
  name = NULL,
  desc = NULL,
  sets = list(),
  aliases = list(),
  mappings = list(),
  parameters = list(),
  variables = list(),
  equations = list()
)
}
\arguments{
\item{desc}{}

\item{sets}{named list of sets}

\item{mappings}{named list of mappings}

\item{parameters}{named list of multimod_parameter objects}

\item{variables}{named list of multimod_variable objects}

\item{equations}{named list of multimod_equation objects}
}
\description{
Create a multimod model object
}
