% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_aliases.R
\name{preferred_dummy_names}
\alias{preferred_dummy_names}
\title{Preferred short names for common sets to avoid conflicts}
\usage{
preferred_dummy_names
}
\description{
Preferred short names for common sets to avoid conflicts
}
\keyword{internal}
