% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{split_bounds_parameter_data}
\alias{split_bounds_parameter_data}
\title{Split bounds-style parameter into low/high tables}
\usage{
split_bounds_parameter_data(ert_param, scenario = NULL)
}
\arguments{
\item{ert_param}{energyRt parameter object of type "bounds"}
}
\value{
List with lo/up data.frames (each containing dims + value)
}
\description{
Split bounds-style parameter into low/high tables
}
\keyword{internal}
