% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_latex.R
\name{write_latex.model}
\alias{write_latex.model}
\title{Write LaTeX representation of a model}
\usage{
\method{write_latex}{model}(
  x,
  file = NULL,
  append = FALSE,
  preamble = NULL,
  ending = NULL,
  title = paste0("Model: ", x$name),
  subtitle = x$desc,
  author = x$authors,
  show_date = TRUE,
  subsection_number = TRUE,
  include_sets = TRUE,
  include_aliases = TRUE,
  include_parameters = TRUE,
  include_variables = TRUE,
  include_equations = TRUE,
  eq_substitute = list(when = "condition"),
  alias_map = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{append}{Logical; if \code{TRUE}, appends to the file instead of overwriting.}

\item{preamble}{Character vector; LaTeX preamble to use. If \code{NULL}, uses the default preamble.}

\item{ending}{Character vector; LaTeX ending to use. If \code{NULL}, uses the default ending.}

\item{subsection_number}{Logical; if \code{TRUE}, includes subsection numbering in the LaTeX output.}

\item{eq_substitute}{Named list; substitutions for specific AST elements in the equation
with \code{ast_where} object to display conditions, indices, etc. below the equation.}

\item{verbose}{Logical; if \code{TRUE}, prints additional information during processing.}

\item{...}{Additional arguments passed to the \code{as_latex} function for rendering the equation.}
}
\description{
Write LaTeX representation of a model
}
